edwards and weierstrass args stored under models
